requirements.txt
openai
python-dotenv
sqlalchemy
duckdb
duckdb-engine
kagglehub # used for sample datasets